app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-ai-documentintelligence==1.0.0b4",
    "pillow>=11.2.1",
    "streamlit>=1.45.1",
]
